+++ src/ModelDataTable.php
@@ -3,12 +3,15 @@
 namespace Exist404\DatatableCruds;
 
 use Exist404\DatatableCruds\Exceptions\ModelIsNotSet;
+use Illuminate\Contracts\Database\Eloquent\Builder;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class ModelDataTable
 {
-    private $defaultRequest = [
+    private array $defaultRequest = [
         'orderBy' => 'created_at',
         'order' => 'desc',
         'page' => 1,
@@ -18,11 +21,12 @@ class ModelDataTable
         'searchBy' => [],
     ];
 
-    private $model;
-    private $tableName;
-    private $request;
+    private Model $model;
+    private string $tableName;
+    private string $primaryKeyName;
+    private object $request;
 
-    public function __construct($model)
+    public function __construct(string $model)
     {
         if (!isset($model)) {
             throw ModelIsNotSet::create();
@@ -30,6 +34,7 @@ class ModelDataTable
         $this->request = (object) array_merge($this->defaultRequest, $_GET);
         $this->model = new $model();
         $this->tableName = $this->model->getTable();
+        $this->primaryKeyName = $this->model->getKeyName();
     }
 
     public function get(): LengthAwarePaginator|Collection
@@ -49,7 +54,7 @@ class ModelDataTable
                     ->paginate($this->request->limit, ['*'], 'page', $this->request->page);
     }
 
-    private function applyOrderRequest()
+    private function applyOrderRequest(): Builder
     {
         if ($this->isRelatedToModel($this->request->orderBy)) {
             return $this->orderByRelated();
@@ -57,7 +62,7 @@ class ModelDataTable
         return $this->model->orderBy($this->request->orderBy, $this->request->order);
     }
 
-    private function applyWithRelations($query)
+    private function applyWithRelations(Builder $query): Builder
     {
         if ($this->request->with) {
             $with = !is_array($this->request->with) ? explode('|', $this->request->with) : $this->request->with;
@@ -68,7 +73,7 @@ class ModelDataTable
         return $query;
     }
 
-    private function applySearch($query)
+    private function applySearch(Builder $query): Builder
     {
         $searchBy = $this->request->searchBy ?: [];
         $searchBy = !is_array($searchBy) ? explode(',', $searchBy) : $searchBy;
@@ -88,23 +93,50 @@ class ModelDataTable
         return $query;
     }
 
-    private function isRelatedToModel($field)
+    private function isRelatedToModel(string $field): bool
     {
         return isset(explode('.', $field)[1]);
     }
 
-    private function orderByRelated()
+    private function orderByRelated(): Builder
     {
-        @list($related, $orderBy) = explode('.', $this->request->orderBy);
-        $relatedModel = $this->model->$related()->getRelated();
-        $relatedTableName = $relatedModel->getTable();
-        $relatedForeign = $relatedModel->getForeignKey();
-        $relatedOwnerKey = $relatedModel->getKeyName();
-        return $this->model->join(
-            $relatedTableName,
-            "{$this->tableName}.$relatedForeign",
-            '=',
-            "$relatedTableName.$relatedOwnerKey"
-        )->orderBy("$relatedTableName.$orderBy", $this->request->order ?? 'desc');
+        @list($relationName, $orderBy) = explode('.', $this->request->orderBy);
+
+        $relatedTableName = $this->relatedTableName($relationName);
+
+        return $this->leftJoinRelation($relationName)
+            ->orderBy("$relatedTableName.$orderBy", $this->request->order ?? 'desc');
+    }
+
+    private function leftJoinRelation(string $relationName): Builder
+    {
+        return $this->model->distinct("{$this->tableName}.{$this->primaryKeyName}")
+            ->leftJoin($this->relatedTableName($relationName), function ($join) use ($relationName) {
+                $relation = $this->model->$relationName();
+                $relatedModel = $this->relatedModel($relationName);
+                $relatedForeign = $relation->getForeignKeyName();
+                $relatedTableName = $relatedModel->getTable();
+                $relatedOwnerKey = $relatedModel->getKeyName();
+                $relatedMorphType = method_exists($relation, "getMorphType") ? $relation->getMorphType() : null;
+                $relatedMorphClass = method_exists($relation, "getMorphClass") ? $relation->getMorphClass() : null;
+                if ($relatedMorphClass) {
+                    $join->on("$relatedTableName.$relatedMorphType", '=', $relatedMorphClass);
+                }
+                if ($relation instanceof BelongsTo) {
+                    $join->on("{$this->tableName}.$relatedForeign", '=', "$relatedTableName.$relatedOwnerKey");
+                } else {
+                    $join->on("$relatedTableName.$relatedForeign", '=', "{$this->tableName}.$relatedOwnerKey");
+                }
+            });
+    }
+
+    private function relatedModel(string $relationName): Model
+    {
+        return $this->model->$relationName()->getModel();
+    }
+
+    private function relatedTableName(string $relationName): string
+    {
+        return $this->relatedModel($relationName)->getTable();
     }
 }