+++ src/routes/web.php
@@ -1,9 +1,12 @@
 <?php
 
-use Exist404\DatatableCruds\Exceptions\ModelIsNotSet;
 use Illuminate\Support\Facades\Route;
 
 Route::get(config('datatablecruds.script_file_url'), function () {
-    $sctiptContent = file_get_contents(__DIR__ . '/../public/js/datatable-cruds.min.js');
-    return response($sctiptContent, 200, ['Content-Type' => 'text/javascript']);
-});
+    $scriptContent = file_get_contents(datatableScriptPath());
+    $response = response($scriptContent, 200, ['Content-Type' => 'text/javascript']);
+    $response->setSharedMaxAge(31536000);
+    $response->setMaxAge(31536000);
+    $response->setExpires(new \DateTime("+1 year"));
+    return $response;
+})->name('datatablecruds.script_file_url');